% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFullTimeseries.R
\name{GetFullTimeseries}
\alias{GetFullTimeseries}
\title{Get full Timeseries}
\usage{
GetFullTimeseries()
}
\value{
A data frame containing Country,Latitude and Longitude, Number of new cases death in each day
}
\description{
This function is a without input function that elicit the Covid data time-series from John Hopkins University data center in github, cleans it and returns it to you.
}
\examples{
  Covid19Map:: GetFullTimeseries()
}
