% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCountryTimeseries.R
\name{PlotCountryTimeseries}
\alias{PlotCountryTimeseries}
\title{Plot Country Time-series}
\usage{
PlotCountryTimeseries(FromDate, ToDate, Country = "Iran", TypeP = "New Death")
}
\arguments{
\item{FromDate}{is the starting date (eg. "1.1.2020").}

\item{TypeP}{is the type of data that you want to show (eg."New Death" or "New Case")}

\item{ToDdate}{is the Finishing date.}

\item{country}{is the country that you want to see the time-series of the data}
}
\value{
A pretty figure
}
\description{
This function is to plot the time-series of a the Covid-19 in a specific country, within specific days and type of data
}
\examples{
Covid19Map::PlotCountryTimeseries(FromDate="2020.2.1",
                                 ToDate="2020.8.1",
                                 Country="Iran",TypeP="New Death")

}
