% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWorldMap.R
\name{PlotWorldMap}
\alias{PlotWorldMap}
\title{Plot World map}
\usage{
PlotWorldMap(DayP = 20, MonthP = 10, YearP = 2020, TypeP = "New Death")
}
\arguments{
\item{DayP}{is the day that you want to plot the data}

\item{MonthP}{is the month that you want to plot the data}

\item{YearP}{is the year that you want to plot the data}

\item{TypeP}{is the type of data that you want to show (eg."New Death" or "New Case")}
}
\value{
A pretty figure
}
\description{
This function is to plot the Covid-19 situation in a specific date, in specific type of data
}
\examples{
Covid19Map::PlotWorldMap(DayP=20,MonthP=10,YearP=2020,TypeP="New Death")
}
